requirements.txt.txt
Flask==2.2.5
torch==2.1.0
torchvision==0.16.0
numpy==1.24.4
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
Pillow==10.0.0
